find position player

+++ rsc/errors_handling.c
@@ -6,7 +6,7 @@
 /*   By: souchen <souchen@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/21 21:24:03 by yismaili          #+#    #+#             */
-/*   Updated: 2022/09/14 11:51:13 by yismaili         ###   ########.fr       */
+/*   Updated: 2022/11/12 13:20:32 by souchen          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -261,7 +261,7 @@ char    **ft_jump_lines(t_struct *cub)
     cmp = 0;
     while(data[i] != NULL)
     {
-        if(ft_strlen(data[i] > max))
+        if((int)ft_strlen(data[i]) > max)
         {
             max = ft_strlen(data[i]);
         }
